added logs for debugging
