+++ next.config.ts
@@ -2,7 +2,13 @@ import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
   // don't transpile `nexlog` - use the package's published builds/exports
-  // transpilePackages: ['nexlog'],
+  webpack: (config, { isServer }) => {
+    if (isServer) {
+      config.externals.push('nexlog');
+    }
+    return config;
+  },
+  transpilePackages: ['nexlog'],
   images: { unoptimized: true },
   trailingSlash: false,
   basePath: process.env.NEXT_PUBLIC_BASE_PATH || '',
