+++ package.json
@@ -22,6 +22,7 @@
     "jwt-decode": "^4.0.0",
     "nexlog": "^5.2.2",
     "next": "^15.2.4",
+    "pino": "^10.1.0",
     "react": "^19.2.0",
     "react-calendar-datetime-picker": "^1.7.5",
     "react-dom": "^19.2.0",
